Rename and reorganize Notification test classes
